FossilOrigin-Name: 0a5ca0f3158a59a4f5bd90bac1f22b685df4e8eb

+++ src/rebuild.c
@@ -176,8 +176,9 @@ static void rebuild_step(int rid, int size, Blob *pBase){
       blob_write_to_file(pUse,zFile);
       free(zFile);
       free(zUuid);
+      blob_reset(pUse);
     }
-    blob_reset(pUse);
+    assert( blob_is_reset(pUse) );
     rebuild_step_done(rid);
   
     /* Call all children recursively */
@@ -342,6 +343,63 @@ int rebuild_db(int randomize, int doOut, int doClustering){
 }
 
 /*
+** Attempt to convert more full-text blobs into delta-blobs for
+** storage efficiency.
+*/
+static void extra_deltification(void){
+  Stmt q;
+  int topid, previd, rid;
+  int prevfnid, fnid;
+  db_begin_transaction();
+  db_prepare(&q,
+     "SELECT rid FROM event, blob"
+     " WHERE blob.rid=event.objid"
+     "   AND event.type='ci'"
+     "   AND NOT EXISTS(SELECT 1 FROM delta WHERE rid=blob.rid)"
+     " ORDER BY event.mtime DESC"
+  );
+  topid = previd = 0;
+  while( db_step(&q)==SQLITE_ROW ){
+    rid = db_column_int(&q, 0);
+    if( topid==0 ){
+      topid = previd = rid;
+    }else{
+      if( content_deltify(rid, previd, 0)==0 && previd!=topid ){
+        content_deltify(rid, topid, 0);
+      }
+      previd = rid;
+    }
+  }
+  db_finalize(&q);
+
+  db_prepare(&q,
+     "SELECT blob.rid, mlink.fnid FROM blob, mlink, plink"
+     " WHERE NOT EXISTS(SELECT 1 FROM delta WHERE rid=blob.rid)"
+     "   AND mlink.fid=blob.rid"
+     "   AND mlink.mid=plink.cid"
+     "   AND plink.cid=mlink.mid"
+     " ORDER BY mlink.fnid, plink.mtime DESC"
+  );
+  prevfnid = 0;
+  while( db_step(&q)==SQLITE_ROW ){
+    rid = db_column_int(&q, 0);
+    fnid = db_column_int(&q, 1);
+    if( prevfnid!=fnid ){
+      prevfnid = fnid;
+      topid = previd = rid;
+    }else{
+      if( content_deltify(rid, previd, 0)==0 && previd!=topid ){
+        content_deltify(rid, topid, 0);
+      }
+      previd = rid;
+    }
+  }
+  db_finalize(&q);
+
+  db_end_transaction(0);
+}
+
+/*
 ** COMMAND:  rebuild
 **
 ** Usage: %fossil rebuild ?REPOSITORY?
@@ -356,6 +414,10 @@ int rebuild_db(int randomize, int doOut, int doClustering){
 **   --force       Force the rebuild to complete even if errors are seen
 **   --randomize   Scan artifacts in a random order
 **   --cluster     Compute clusters for unclustered artifacts
+**   --pagesize N  Set the database pagesize to N. (512..65536 and power of 2)
+**   --wal         Set Write-Ahead-Log journalling mode on the database
+**   --compress    Strive to make the database as small as possible
+**   --vacuum      Run VACUUM on the database after rebuilding
 */
 void rebuild_database(void){
   int forceFlag;
@@ -363,11 +425,28 @@ void rebuild_database(void){
   int errCnt;
   int omitVerify;
   int doClustering;
+  const char *zPagesize;
+  int newPagesize = 0;
+  int activateWal;
+  int runVacuum;
+  int runCompress;
 
   omitVerify = find_option("noverify",0,0)!=0;
   forceFlag = find_option("force","f",0)!=0;
   randomizeFlag = find_option("randomize", 0, 0)!=0;
   doClustering = find_option("cluster", 0, 0)!=0;
+  runVacuum = find_option("vacuum",0,0)!=0;
+  runCompress = find_option("compress",0,0)!=0;
+  zPagesize = find_option("pagesize",0,1);
+  if( zPagesize ){
+    newPagesize = atoi(zPagesize);
+    if( newPagesize<512 || newPagesize>65536
+        || (newPagesize&(newPagesize-1))!=0
+    ){
+      fossil_fatal("page size must be a power of two between 512 and 65536");
+    }
+  }
+  activateWal = find_option("wal",0,0)!=0;
   if( g.argc==3 ){
     db_open_repository(g.argv[2]);
   }else{
@@ -391,8 +470,28 @@ void rebuild_database(void){
             errCnt);
     db_end_transaction(1);
   }else{
+    if( runCompress ){
+      printf("Extra delta compression... "); fflush(stdout);
+      extra_deltification();
+      runVacuum = 1;
+    }
     if( omitVerify ) verify_cancel();
     db_end_transaction(0);
+    if( runCompress ) printf("done\n");
+    db_close(0);
+    db_open_repository(g.zRepositoryName);
+    if( newPagesize ){
+      db_multi_exec("PRAGMA page_size=%d", newPagesize);
+      runVacuum = 1;
+    }
+    if( runVacuum ){
+      printf("Vacuuming the database... "); fflush(stdout);
+      db_multi_exec("VACUUM");
+      printf("done\n");
+    }
+    if( activateWal ){
+      db_multi_exec("PRAGMA journal_mode=WAL;");
+    }
   }
 }
 
@@ -509,7 +608,7 @@ void test_clusters_cmd(void){
 
 /*
 ** COMMAND: scrub
-** %fossil scrub [--verily] [--force] [REPOSITORY]
+** %fossil scrub [--verily] [--force] [--private] [REPOSITORY]
 **
 ** The command removes sensitive information (such as passwords) from a
 ** repository so that the respository can be sent to an untrusted reader.
@@ -517,7 +616,8 @@ void test_clusters_cmd(void){
 ** By default, only passwords are removed.  However, if the --verily option
 ** is added, then private branches, concealed email addresses, IP
 ** addresses of correspondents, and similar privacy-sensitive fields
-** are also purged.
+** are also purged.  If the --private option is used, then only private
+** branches are removed and all other information is left intact.
 **
 ** This command permanently deletes the scrubbed information.  The effects
 ** of this command are irreversible.  Use with caution.
@@ -528,6 +628,7 @@ void test_clusters_cmd(void){
 void scrub_cmd(void){
   int bVerily = find_option("verily",0,0)!=0;
   int bForce = find_option("force", "f", 0)!=0;
+  int privateOnly = find_option("private",0,0)!=0;
   int bNeedRebuild = 0;
   if( g.argc!=2 && g.argc!=3 ) usage("?REPOSITORY?");
   if( g.argc==2 ){
@@ -538,26 +639,33 @@ void scrub_cmd(void){
   if( !bForce ){
     Blob ans;
     blob_zero(&ans);
-    prompt_user("Scrubbing the repository will permanently remove user\n"
-                "passwords and other information. Changes cannot be undone.\n"
-                "Continue (y/N)? ", &ans);
+    prompt_user("Scrubbing the repository will permanently information.\n"
+                "Changes cannot be undone.  Continue (y/N)? ", &ans);
     if( blob_str(&ans)[0]!='y' ){
       fossil_exit(1);
     }
   }
   db_begin_transaction();
-  db_multi_exec(
-    "UPDATE user SET pw='';"
-    "DELETE FROM config WHERE name GLOB 'last-sync-*';"
-  );
-  if( bVerily ){
+  if( privateOnly || bVerily ){
     bNeedRebuild = db_exists("SELECT 1 FROM private");
     db_multi_exec(
-      "DELETE FROM concealed;"
-      "UPDATE rcvfrom SET ipaddr='unknown';"
-      "UPDATE user SET photo=NULL, info='';"
-      "INSERT INTO shun SELECT uuid FROM blob WHERE rid IN private;"
+      "DELETE FROM blob WHERE rid IN private;"
+      "DELETE FROM delta WHERE rid IN private;"
+      "DELETE FROM private;"
     );
+  }
+  if( !privateOnly ){
+    db_multi_exec(
+      "UPDATE user SET pw='';"
+      "DELETE FROM config WHERE name GLOB 'last-sync-*';"
+    );
+    if( bVerily ){
+      db_multi_exec(
+        "DELETE FROM concealed;"
+        "UPDATE rcvfrom SET ipaddr='unknown';"
+        "UPDATE user SET photo=NULL, info='';"
+      );
+    }
   }
   if( !bNeedRebuild ){
     db_end_transaction(0);
@@ -597,13 +705,14 @@ void recon_read_dir(char *zPath){
         fossil_panic("some unknown error occurred while reading \"%s\"", 
                      blob_str(&path));
       }
-      content_put(&aContent, 0, 0, 0);
+      content_put(&aContent);
       blob_reset(&path);
       blob_reset(&aContent);
       free(zSubpath);
       printf("\r%d", ++nFileRead);
       fflush(stdout);
     }
+    closedir(d);
   }else {
     fossil_panic("encountered error %d while trying to open \"%s\".",
                   errno, g.argv[3]);
